requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
chromedriver==2.24.1
click==6.7
idna==2.6                 # via requests
pymongo==3.6.0
python-dateutil==2.6.1
requests==2.18.4
selenium==3.8.0
six==1.11.0               # via python-dateutil
urllib3==1.22             # via requests
